Add support for APF related Wi-Fi HAL APIs in cuttlefish
This commit introduces initial support for the wifi_set_packet_filter()
and wifi_read_packet_filter() Wi-Fi HAL APIs in cuttlefish. This paves
the way for full APF (Android Packet Filtering) functionality. The next
step is to determine the method for running the APF program to in the
actual interface.

Test: TH
Change-Id: I6701f65b40338c436af87c405de6b1a57511114b