Fix boot issues in slim image.
Previously the slim image used the 'phone' config
as a baseline. This is not appropriate for the slim image
since it includes aosp_product and handheld_system_ext and caused
boot failures.

This commit creates a new base_handheld config that doesn't
contain  aosp_product and handheld_system_ext, that a new slim_handheld
config can inherit and customize.

Bug: 341171095
Test: boot
Change-Id: Id251df32ce0a52a5c93f1177267c72a6035d5364