[Thread] add Thread HAL to goldfish
This commit adds the Thread HAL to goldfish. It's similar to the same
Thread HAL added to Cuttlefish devices but this is also useful for app
development and testing.

For background, Thread is a new wireless technology added to Android. It
requires new 802.15.4 radio chip to operate, but we have a simulation
radio chip device for goldfish and cuttlefish.

Bug: 356079409
Test: verified that "adb shell cmd thread_network join" works
Change-Id: I414d70238b807608f86cb2d06f844ff31993c1ef