goldfish: Rename kernel cmdline "page_shift"
androidboot.page_shift was originally named this way since init
creates ro sys props for all the androidboot parameters. But it
really is a kernel command line option not a bootconfig option.
The sys prop is not currently used in the platform.

Rename 'androidboot.page_shift' to just 'page_shift' as VTS will
fail if there are androidboot* options found in the kernel command
line and not in bootconfig.

Bug: 356439363
Test: atest VtsBootconfigTest
Change-Id: I4c917567f40c85937718cabf26b974eb89a3f993
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>