Add kernel command line files
we need a way to configure the Android kernel in
a system image specific way. This is currently
done on the host side, but the host does not have
all the information it needs.

With these files added, some host side configuration
is no longer necessary and will be removed if this
artifact is found in the system image.

Bug: 356858341
Bug: 356176653
Test: manual
Change-Id: I82c66d0663471f197ba908432a1bac712804a5fa
Signed-off-by: Roman Kiryanov <rkir@google.com>